moves filters into separate file and adds invalid filter check to boolean
